Check the classes of a data frame and allow for multiple different classes for certain columns

This function checks the classes of a data frame and allows for certain columns to have multiple allowed classes.

Usage

check_df(df, class_list)

Arguments

df

The data frame to check

class_list

A named list that specifies the expected classes for each column

The class_list argument should be a named list, where the names correspond to the column names in df. Each entry in the list should be a character vector of the allowed classes for the corresponding column. For example, class_list = list(col1 = "numeric", col2 = c("character", "factor"), col3 = "logical") specifies that col1 must be of class "numeric", col2 must be of class "character" or "factor", and col3 must be of class "logical".

Value

The checked data frame

Examples

if (FALSE) { # \dontrun{
df <- data.frame(
  col1 = c(1, 2, 3),
  col2 = c("a", "b", "c"),
  col3 = c(TRUE, FALSE, TRUE),
  stringsAsFactors = FALSE
)

class_list <- list(
  col1 = "numeric",
  col2 = c("character", "factor"),
  col3 = "logical"
)

check_df(df, class_list)
} # }