Get Draft Group Info

Fetch info for a specific draft group

Usage

dk_get_draft_group_info(
  draft_group_id = NULL,
  contest_key = NULL,
  output = c("cleaned_json", "json", "response", "request"),
  process_args = NULL,
  ...
)

Arguments

draft_group_id

Sequence of digits that correspond to a draft table/group. If draft_group_id and contest_key are both passed, contest_key is ignored.

contest_key

The sequence of digits that correspond to a specific contest. This can be found by examining the URL of a contest page. For example: https://www.draftkings.com/draft/contest/133645678#. Here the contest ID is 133645678.

output

One of "cleaned_json" (the default), "json", "response", or "request". If "cleaned_json" then tidyjson::spread_all() is used to parse the JSON body,

process_args

Optional list of arguments passed to dk_req_process

...

Arguments passed on to dk_req

query_params

A list of query parameters passed to [httr2::req_url_query].

proxy_args

List of arguments to httr2::req_proxy

curl_options

List of arguments to httr2::req_options

throttle_rate

List of arguments to httr2::req_throttle

headers

List of arguments to httr2::req_headers()

paths

List of arguments to httr2::req_url_path_append()

error_handling_options

List of arguments to httr2::req_error()

retry_options

List of arguments passed to httr2::req_retry()

base_url

Base URL for request.

method

Custom HTTP method

Examples

if (FALSE) { # \dontrun{
  dk_get_draft_group_info(75284)

} # }