Get Entry Details

Fetch details for entries in a contest, including the drafted roster for each entry, stats for each player in an entry roster, and the fantasy points associated to each stat.

Usage

dk_get_entries(
  draft_group_id,
  entry_keys,
  iv = Sys.getenv("DK_IV"),
  jwe = Sys.getenv("DK_JWE"),
  output = c("cleaned_json", "json", "response", "request"),
  process_args = NULL,
  ...
)

Arguments

draft_group_id

Sequence of digits that correspond to a draft table/group. If draft_group_id and contest_key are both passed, contest_key is ignored.

entry_keys

Vector of numeric (or character) keys that correspond to a specific entry in a specific contest. See output from dk_get_leaderboard().

iv

Character string. The 'iv' cookie value. If not provided, it will be retrieved from the DK_IV environment variable.

jwe

Character string. The 'jwe' cookie value. If not provided, it will be retrieved from the DK_JWE environment variable.

output

One of "cleaned_json" (the default), "json", "response", or "request". If "cleaned_json" then tidyjson::spread_all() is used to parse the JSON body,

process_args

Optional list of arguments passed to dk_req_process

...

Arguments passed on to dk_req

query_params

A list of query parameters passed to [httr2::req_url_query].

proxy_args

List of arguments to httr2::req_proxy

curl_options

List of arguments to httr2::req_options

throttle_rate

List of arguments to httr2::req_throttle

headers

List of arguments to httr2::req_headers()

paths

List of arguments to httr2::req_url_path_append()

error_handling_options

List of arguments to httr2::req_error()

retry_options

List of arguments passed to httr2::req_retry()

base_url

Base URL for request.

method

Custom HTTP method

Examples

  if (FALSE) { # \dontrun{
    dk_get_entries(draft_group_id = 80584, entry_keys = c(3618408508, 3618897002))
  } # }