Optimize Lineup

Get optimal players based on player pool, projected fantasy points, and rules defined by a schematic. See dk_prepare_schematic().

Determine the optimal lineup of players based on the list of available players, their salaries, and their projected fantasy points for a showdown captain mode game type.

Determine the optimal lineup of players based on the list of available players, their salaries, and their projected fantasy points for a classic game type.

Usage

dk_optimize_lineup(schematic, max_points, solver = "glpk", ...)

# S3 method for class 'showdown_captain_mode'
dk_optimize_lineup(schematic, max_points = NULL, solver = "glpk", ...)

# S3 method for class 'classic'
dk_optimize_lineup(schematic, max_points = NULL, solver = "glpk", ...)

Arguments

schematic

Output from dk_prepare_schematic() which includes information needed for optimization; including player info such as, news status, salary, and projected fantasy points.

max_points

Optional upper threshold of points. This is used primarily to get many successive lineups in dk_get_optimal_lineups()

solver

the 'ROI' solver name (character vector of length 1)

...

Other arguments passed to optimization method.