Write Lineups to CSV

Create CSV of optimal lineups that can be uploaded to https://www.draftkings.com/lineup/upload.

Given the output from dk_get_optimal_lineups(), create a CSV of Showdown Captain Mode lineups formatted for uplaod to https://www.draftkings.com/lineup/upload.

Usage

dk_write_csv(optimal_lineups, file, ...)

# S3 method for class 'showdown_captain_mode_multiple_solutions'
dk_write_csv(optimal_lineups, file = "showdown_captain_mode_lineups.csv", ...)

Arguments

optimal_lineups

Output from dk_get_optimal_lineups()

file

Path and name of the CSV file to create

...

Other args passed to utils::write.csv()